+++ index.html
@@ -814,5 +814,6 @@
             <div class="ct-chart ct-perfect-fourth" id = "ffff"></div>
         </div>
         <script src="index.js"></script>
+        <script src="Chart.js"></script>
     </body>
 </html>
