+++ index.js
@@ -1,5 +1,3 @@
-import {chart} from './Chart';
-
 const vanyacha = document.getElementById('content');
 const buttom_top1 = document.getElementById('buttom_slider1');
 const buttom_top2 = document.getElementById('buttom_slider2');
@@ -16,121 +14,10 @@ const slidContentThree = document.getElementById("slid_content_three");
 const OneSlid = document.getElementById("one_slid");
 const TwoSlid = document.getElementById("two_slid");
 const ThreeSlid = document.getElementById("three_slid");
-const chart1 = document.getElementById("c1");
-const statusOne = document.getElementById("Status_Favourites");
-const statusTwo = document.getElementById("Status_Posts_Last_24_Hours");
-const statusThree = document.getElementById("Status_Total_Posts");
-const statusFour= document.getElementById("Status_Campaigns");
-const statusFive = document.getElementById("Status_Amazing_Features");
 const DisabledButton = document.getElementById("check_in_button");
-const OffButton = document.getElementById("off_button");
-const body = document.getElementById("body");
-const overflow = document.createElement('div');
-const UF = [
-    [1,3,5,6,8,9,5,3,2,4,6,7],
-    [1,3,5,6,8,9,5,3,2,4,6,7]
-]
-const PLH = [
-    [8,2,8,10,2,4,1,5,9,2,4,5],
-    [5,3,5,2,9,4,5,8,6,10,3,4]
-]
-const TP = [
-    [4,7,3,8,8,7,3,6,1,5,9,10],
-    [5,9,2,3,9,4,8,6,8,3,2,5]
-]
-const Campaigns = [
-    [4,1,9,10,3,8,2,3,8,6,3,5],
-    [6,3,5,6,8,9,5,3,2,4,6,7]
-]
-const AF = [
-    [10,2,7,10,8,9,5,3,2,4,6,7],
-    [7,3,5,6,8,9,5,3,2,4,6,7]
-]
-
-let tittleChart = document.getElementById("tittle_chart");
 let Input = document.getElementsByClassName('info_check_in');
 let inputNam = document.getElementById("Name");
 let inputEm = document.getElementById("Email");
-
-chart1.style.display = 'none';
-
-function ofChart () {
-    chart1.style.display = 'none';
-    overflow.style.display = 'none';
-    body.style.position = '';
-    body.style.overflow = 'visible';
-}
-
-OffButton.addEventListener('click', () => {
-    ofChart()
-})
-
-function generateData (type) {
-    switch (type) {
-        case 'one': type = UF;
-        break;
-
-        case 'two': type = PLH;
-        break;
-
-        case 'three': type = TP;
-        break;
-
-        case 'foure': type = Campaigns;
-        break;
-
-        case 'five': type = AF;
-        break;
-    }
-    chart1.style.display = "inline-block";
-    chartOnStyle();
-    modal(); 
-    return {
-        labels: ['Jan', 'Feb', 'Mar', 'Apr', 'Mai', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'],
-        series: type,
-    };
-} 
-
-function modal() {
-    overflow.className = 'overflow';
-    body.style.position = 'relative';
-    document.body.appendChild(overflow);
-    overflow.style.display = 'block';
-    overflow.addEventListener('click', () => {
-        ofChart()
-    })
-}
-
-function chartOnStyle () {
-    body.style.overflow = 'hidden';
-    body.style.zIndex = '100';
-}
-
-
-statusOne.addEventListener('click', () => {
-    chart(generateData('one'));
-    tittleChart.innerHTML = 'User Favourites';
-})
-
-statusTwo.addEventListener('click', () => {
-    chart(generateData('two'));
-    tittleChart.innerHTML = 'Posts Last 24 Hours';
-})
-
-statusThree.addEventListener('click', () => {
-    chart(generateData('three'));
-    tittleChart.innerHTML = 'Total Posts';
-})
-
-statusFour.addEventListener('click', () => {
-    chart(generateData('foure'));
-    tittleChart.innerHTML = 'Campaigns';
-})
-
-statusFive.addEventListener('click', () => {
-    chart(generateData('five'));
-    tittleChart.innerHTML = 'Amazing Features';
-})
 
 buttom_top1.addEventListener('click', () => {
     buttom_slider_c1.style.backgroundColor = 'e74c3c';
@@ -235,13 +122,6 @@ function InputValidator (e) {
 
 inputNam.addEventListener('input', () => {InputValidator(Input)})
 inputEm.addEventListener('input', () => {InputValidator(Input)})
-
-document.addEventListener('keydown', function(event) {
-    if(event.keyCode === 27) {
-        ofChart()
-    }
-})
-
